requirement.txt
# ============ 深度学习框架 ============
torch>=1.12.0
torchvision>=0.13.0

# ============ GUI框架 ============
PyQt5>=5.15.0

# ============ 图像处理 ============
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.19.0

# ============ 机器学习 ============
scikit-learn>=1.0.0
xgboost>=1.6.0
catboost>=1.1.0
lightgbm>=3.3.0
joblib>=1.1.0

# ============ 视觉模型库 ============
timm>=0.6.0  # Vision Transformer等模型
grad-cam

# ============ 科学计算 ============
numpy>=1.21.0
pyqt5-tools